requirements.txt
kafka-python-ng==2.2.2
pyspark==3.5.0
websockets
asyncio
delta-spark==3.0.0
streamlit
duckdb
pydantic
requests
pytest
python-dotenv
plotly

